requirements.txt
django>=1.8,<1.9
djangorestframework>=3.3,<3.4
django-extensions>=1.5,<1.6
django-grappelli>=2.7,<2.8
django-mptt>=0.7,<0.8
django-tables2>=1.0,<1.1
django-registration>=2.0,<2.1
django-braces>=1.8,<1.9
